+++ src/TapkuLibrary/NSString+TKCategory.m
@@ -67,8 +67,8 @@
 - (NSString *) escapeHTML{
 	NSMutableString *s = [NSMutableString string];
 	
-	int start = 0;
-	int len = [self length];
+	NSInteger start = 0;
+	NSInteger len = [self length];
 	NSCharacterSet *chs = [NSCharacterSet characterSetWithCharactersInString:@"<>&\""];
 	
 	while (start < len) {
