+++ src/TapkuLibrary/UIColor+TKCategory.m
@@ -49,9 +49,9 @@
 
 + (UIColor*) randomColor{
 	
-	int r = arc4random() % 255;
-	int g = arc4random() % 255;
-	int b = arc4random() % 255;
+	NSInteger r = arc4random() % 255;
+	NSInteger g = arc4random() % 255;
+	NSInteger b = arc4random() % 255;
 	return [UIColor colorWithRed:r/255.0 green:g/255.0 blue:b/255.0 alpha:1];
 	
 }
